// src/api-client/models/MainPostPreview.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Mantle API
 * A static API for blogs that transforms markdown files with frontmatter into JSON endpoints
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { MainFrontMatter } from './MainFrontMatter';
import {
    MainFrontMatterFromJSON,
    MainFrontMatterFromJSONTyped,
    MainFrontMatterToJSON,
    MainFrontMatterToJSONTyped,
} from './MainFrontMatter';

/**
 * Post preview containing frontmatter, excerpt, and reading time
 * @export
 * @interface MainPostPreview
 */
export interface MainPostPreview {
    /**
     * 
     * @type {string}
     * @memberof MainPostPreview
     */
    excerpt?: string;
    /**
     * 
     * @type {MainFrontMatter}
     * @memberof MainPostPreview
     */
    frontmatter?: MainFrontMatter;
    /**
     * 
     * @type {number}
     * @memberof MainPostPreview
     */
    readingTime?: number;
}

/**
 * Check if a given object implements the MainPostPreview interface.
 */
export function instanceOfMainPostPreview(value: object): value is MainPostPreview {
    return true;
}

export function MainPostPreviewFromJSON(json: any): MainPostPreview {
    return MainPostPreviewFromJSONTyped(json, false);
}

export function MainPostPreviewFromJSONTyped(json: any, ignoreDiscriminator: boolean): MainPostPreview {
    if (json == null) {
        return json;
    }
    return {
        
        'excerpt': json['excerpt'] == null ? undefined : json['excerpt'],
        'frontmatter': json['frontmatter'] == null ? undefined : MainFrontMatterFromJSON(json['frontmatter']),
        'readingTime': json['readingTime'] == null ? undefined : json['readingTime'],
    };
}

export function MainPostPreviewToJSON(json: any): MainPostPreview {
    return MainPostPreviewToJSONTyped(json, false);
}

export function MainPostPreviewToJSONTyped(value?: MainPostPreview | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'excerpt': value['excerpt'],
        'frontmatter': MainFrontMatterToJSON(value['frontmatter']),
        'readingTime': value['readingTime'],
    };
}